requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
frozenlist==1.6.0
google==3.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.25
langchain-text-splitters==0.3.8
langsmith==0.4.1
marshmallow==3.26.1
motor==3.7.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
openai==1.90.0
orjson==3.10.18
packaging==24.2
passlib==1.7.4
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.1
py==1.11.0
pyasn1==0.6.1
pycryptodome==3.9.9
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.13.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
volcengine==1.0.189
yarl==1.20.0
zstandard==0.23.0
